+++ client/src/components/Today/style/Activity.style.ts
@@ -1,4 +1,4 @@
-import { rowHeight } from "@/components/Today/style/Row.style";
+import { rowHeight } from "@/components/Today/style/TimelineRow.style";
 import { styled } from "styled-components";
 import S from "./Today.style";
